requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
apache-airflow==2.6.3
apache-airflow-providers-common-sql==1.6.0
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.4.2
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
cachelib==0.9.0
cachetools==5.3.1
catalogue==2.0.8
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
clickclick==20.10.2
cmake==3.26.4
colorama==0.4.6
colorlog==4.8.0
command-not-found==0.3
confection==0.1.0
ConfigUpdater==3.1.1
connexion==2.14.2
contourpy==1.1.0
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.18
Deprecated==1.2.14
dill==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.41.0
frozenlist==1.3.3
fsspec==2023.6.0
gensim==4.3.1
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-oauth2-tool==0.0.3
google-re2==1.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==4.6.4
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
keyring==23.5.0
kiwisolver==1.4.4
langcodes==3.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lit==16.0.6
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
netifaces==0.11.0
networkx==3.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.9.0
pathy==0.10.2
pendulum==2.1.2
Pillow==10.0.0
pluggy==1.2.0
preshed==3.0.8
prison==0.2.1
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.7.0
pyparsing==2.4.7
pyphen==0.14.0
python-apt==2.4.0+ubuntu1
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.4.2
rich-argparse==1.2.0
rpds-py==0.8.8
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.1
sentimentpl==0.0.6
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
srsly==2.4.6
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
textstat==0.7.3
thinc==8.1.10
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.16
wadllib==1.3.6
wasabi==1.1.2
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==1.0.0
